refactor: Define new `media` submodule in `native-utils` package
