+++ mobile/modules/native-utils/src/NativeUtils.ts
@@ -1,0 +1,2 @@
+import { requireNativeModule } from "expo";
+export default requireNativeModule("NativeUtils");
