+++ mobile/modules/native-utils/src/NativeUtilsModule.ts
@@ -1,54 +1,9 @@
 import { NativeModule, requireNativeModule } from "expo";
 
-interface AssetsOptions {
-  first: number;
-  after?: number;
-  /** Return results from assets with the following ids. */
-  fromIds?: string[];
-  /** Whether metadata should also be returned with the query. Only effective on Android 11+. */
-  returnWithMetadata?: boolean;
-}
-
-export type Asset = {
-  id: string;
-  filename: string;
-  uri: string;
-  mimeType: string;
-  modificationTime: number;
-  duration: number;
-  fileSize: number;
-
-  /** Only available on Android 11+. */
-  metadata: MusicMetadataAsset | null;
-};
-
-export type MusicMetadataAsset = {
-  title: string | null;
-  album: string | null;
-  albumArtist: string | null;
-  artist: string | null;
-  genre: string | null;
-  year: number | null;
-  discNumber: number | null;
-  trackNumber: number | null;
-  bitrate: number | null;
-
-  /** Only available on Android 16+. */
-  sampleRate: number | null;
-};
-
-export type MusicAssetResult = {
-  assets: Asset[];
-  hasNextPage: boolean;
-  endCursor: number;
-  totalCount: number;
-};
-
 declare class NativeUtilsModule extends NativeModule {
   isSystemDarkMode: boolean;
   launchAppViaIntent(): void;
   saveBundledAssetToURI(assetName: string, toUri: string): Promise<void>;
-  getMusicAssets(options: AssetsOptions): Promise<MusicAssetResult>;
 }
 
 const nativeModule = requireNativeModule<NativeUtilsModule>("NativeUtils");
@@ -66,10 +21,4 @@ export function launchAppViaIntent() {
 export async function saveBundledAssetToURI(assetName: string, toUri: string) {
   if (__DEV__) return;
   return nativeModule.saveBundledAssetToURI(assetName, toUri);
-}
-
-export async function getMusicAssets(
-  options: AssetsOptions,
-): Promise<MusicAssetResult> {
-  return nativeModule.getMusicAssets(options);
 }
