+++ mobile/modules/native-utils/src/media/index.ts
@@ -1,0 +1,2 @@
+export * from "./core";
+export type * from "./types";
